refactor: merge workflows module back to gale
