+++ daggerverse/gale/dagger.json
@@ -6,7 +6,6 @@
     "../actions/artifacts",
     "../actions/artifactcache",
     "../ghx",
-    "../workflows",
     "github.com/aweris/daggerverse/docker@969bc30a84b48f76a28afab3986144d2e47469df"
   ]
 }
