--- daggerverse/workflows/.gitignore
@@ -1,3 +1,0 @@
-/dagger.gen.go
-/internal/querybuilder/
-/querybuilder/
