--- daggerverse/workflows/dagger.json
@@ -1,5 +1,0 @@
-{
-  "name": "workflows",
-  "root": "../..",
-  "sdk": "go"
-}
