+++ go.work
@@ -14,6 +14,5 @@ use (
 	daggerverse/gha/aquasecurity/trivy-action
 	daggerverse/gha/trufflesecurity/trufflehog
 	daggerverse/ghx
-	daggerverse/workflows
 	ghx
 )
